Revert "App-related actions support tag authentication."
This reverts commit c8b1514beabaa2b58ceca1480224bb2d6ff59cf8.

+++ services/cloudapi/create_app.go
@@ -85,7 +85,7 @@ type CreateAppRequest struct {
 type CreateAppResponse struct {
 	*responses.BaseResponse
 	RequestId string `json:"RequestId" xml:"RequestId"`
-	AppId     int64  `json:"AppId" xml:"AppId"`
+	AppId     int    `json:"AppId" xml:"AppId"`
 }
 
 // CreateCreateAppRequest creates a request to invoke CreateApp API
